Fix: Fixed update.sh not working as intended.

+++ update.sh
@@ -2,12 +2,16 @@
 # Radmon Updater
 
 echo "[*] Pulling latest changes from GitHub..."
+git config core.fileMode false
 git pull origin main
+
 
 echo "[*] Updating dependencies in venv..."
 ./venv/bin/pip install --upgrade git+https://github.com/Liptoon/radiacode.git
 
-./venv/bin/pip install -r requirements.txt
+if [ -f "requirements.txt" ]; then
+    ./venv/bin/pip install -r requirements.txt
+fi
 
 chmod +x radmon.py
